Refactor disable() to undo rather than reinitialize
Previously, disable() was destroying the whole indicatorn, reinitialized it and
put it back in the panel. This had several problems, like leaking signals,
positioning errors and bad interference with other extensions.

This has been fixed to retrace all enable() steps and undo them, restoring the
indicator in its previous state. The extra monkey-patched functions remain, as
these have been moved to init(), but this is hopefully not a big burden to
carry when having the extension installed.

Many thanks to Giovanni Campagna for the suggestion.