Prefer the primary device over the first
The code as it was called GetDevices and looped over the devices until either a
primary device or a battery was found. That worked out perfectly on my laptop,
but was failing on more complicated setups, such as those with multiple
batteries.

Unfortunately, we can't rely on always having a primary device (on my laptop,
there is none when the charger is plugged in).

Hence, this keeps the GetDevices() loop but modifies it to always prefer the
primary device if such exists and use the first device only as a fallback of
last resort.

Many thanks to Aurelien Jacobs for the original attempt and valuable feedback.